Selenium_FrameWork_Training - Working Project For Selenium Training
17-March-2018

Selenium_FrameWork_Training - Project is added
It is ready to run from Jenkins. Jenkins Job was configured and the name is : Selenium_FrameWork_Training
It is running as testNG and Maven. It will take screenshot for test method failures. It is a good one for selenium training
Latest Drivers are available at D:\Selenium_Drivers